GRHB-264: * move module to module folder
